Refactor: server/api/health -> server/api/service
