+++ server/api/service/ping/controller.js
